+++ pkg/R/graphics.R
@@ -268,6 +268,10 @@ plot.adaptProj <- function(x, repl = 1, agg_only = FALSE, auto_ylim = TRUE,
 #' @param auto_axis A logical value indicating if the axis labels should be set
 #' as the first variable variable in the input trait axis in the given
 #' \code{adaptInv} object for PIPs. Defaults to \code{FALSE}.
+#' @param auto_axis_ticks A single integer value giving the number of ticks to
+#' include on both axes of a PIP or on the x axis of an elasticity plot, if
+#' \code{auto_axis = TRUE}. Defaults to \code{5}, or to the number of variants
+#' in the used trait axis, whichever is smaller.
 #' @param ... Other parameters used by functions \code{plot.default()}.
 #' 
 #' @return A contour plot showing the overall fitness dynamics of the invader
@@ -352,13 +356,15 @@ plot.adaptProj <- function(x, repl = 1, agg_only = FALSE, auto_ylim = TRUE,
 plot.adaptInv <- function(x, res_variant = 1, inv_variant = 2, repl = 1,
   pip = TRUE, elast = FALSE, run = 1, filled = TRUE, plot.title, plot.axes,
   axes = TRUE, frame.plot = TRUE, auto_ylim = TRUE, auto_col = TRUE,
-  auto_lty = TRUE, auto_title = FALSE, auto_axis = FALSE, ...) {
+  auto_lty = TRUE, auto_title = FALSE, auto_axis = FALSE,
+  auto_axis_ticks = 5, ...) {
   
   asp <- NA
   las <- 1
   xaxs <- yaxs <- "i"
   xlab <- ylab <- used_trait <- used_trait_res <- used_trait_inv <- NULL
   used_trait_axis <- used_trait_spectrum <- NULL
+  elast_args <- list()
   chosen_colours <- c("white", "darkgrey")
   
   if (!axes) frame.plot <- FALSE
@@ -465,6 +471,10 @@ plot.adaptInv <- function(x, res_variant = 1, inv_variant = 2, repl = 1,
     }
   }
   
+  if (!is.element("trait_axis", names(x))) {
+    stop("Argument x does not appear to be an adaptInv object.", call. = FALSE)
+  }
+  
   if (pip) {
     if (is.element("las", found_terms)) las <- further_args$las
     if (is.element("xaxs", found_terms)) xaxs <- further_args$xaxs
@@ -508,13 +518,9 @@ plot.adaptInv <- function(x, res_variant = 1, inv_variant = 2, repl = 1,
       stop("Replicate entered in argument repl could not be found.", call. = FALSE)
     }
     
-    if (!is.element("trait_axis", names(x))) {
-      stop("Argument x does not appear to be an adaptInv object.", call. = FALSE)
-    }
-    
     used_trait_axis <- x$trait_axis
     
-    if (auto_axis[1] == TRUE) {
+    if (auto_axis[1]) {
       var_vector <- c(stats::var(used_trait_axis$givenrate), stats::var(used_trait_axis$offset),
         stats::var(used_trait_axis$multiplier), NA, NA, stats::var(used_trait_axis$surv_dev),
         stats::var(used_trait_axis$obs_dev), stats::var(used_trait_axis$size_dev),
@@ -549,7 +555,7 @@ plot.adaptInv <- function(x, res_variant = 1, inv_variant = 2, repl = 1,
         used_trait_spectrum <- used_trait_axis[, best_candidate + 11]
       }
     }
-    
+      
     res_column <- which(names(x$fitness) == fit_var_name_res)[1]
     resfit_column <- which(names(x$fitness) == fit_var_name_fitres)[1]
     inv_column <- which(names(x$fitness) == fit_var_name_inv)[1]
@@ -607,11 +613,22 @@ plot.adaptInv <- function(x, res_variant = 1, inv_variant = 2, repl = 1,
       if (axes) {
         if (auto_axis[1]) {
           title(main = "", xlab = used_trait_res, ylab = used_trait_inv)
-          range_of_spectrum <- seq(from = min(used_trait_spectrum),
-            to = max(used_trait_spectrum), length.out = length(unique_variants))
           
-          Axis(unique_variants, unique_variants, side = 1, labels = range_of_spectrum)
-          Axis(unique_variants, unique_variants, side = 2, labels = range_of_spectrum)
+          found_length_of_range <- length(used_trait_spectrum)
+          
+          if (found_length_of_range < auto_axis_ticks) auto_axis_ticks <- found_length_of_range
+          
+          range_of_spectrum <- round(seq(from = min(used_trait_spectrum),
+            to = max(used_trait_spectrum), length.out = auto_axis_ticks),
+            digits = 3)
+          
+          at_range_of_spectrum <- seq(from = 1, to = found_length_of_range,
+            length.out = auto_axis_ticks)
+          
+          Axis(unique_variants, unique_variants, side = 1,
+            at = at_range_of_spectrum, labels = range_of_spectrum)
+          Axis(unique_variants, unique_variants, side = 2,
+            at = at_range_of_spectrum, labels = range_of_spectrum)
         } else {
           title(main = "", xlab = "", ylab = "")
           Axis(unique_variants, side = 1)
@@ -661,7 +678,7 @@ plot.adaptInv <- function(x, res_variant = 1, inv_variant = 2, repl = 1,
     if (!is.element("ylab", names(further_args))) {
       further_args$ylab <- "Invader fitness"
     }
-    if (!is.element("xlab", names(further_args))) {
+    if (!is.element("xlab", names(further_args)) & !auto_axis[1]) {
       further_args$xlab <- "Variant"
     }
     
@@ -689,10 +706,66 @@ plot.adaptInv <- function(x, res_variant = 1, inv_variant = 2, repl = 1,
       further_args$lty <- used_lty
     }
     
+    used_trait_axis <- x$trait_axis
+    
+    if (auto_axis[1]) {
+      var_vector <- c(stats::var(used_trait_axis$givenrate), stats::var(used_trait_axis$offset),
+        stats::var(used_trait_axis$multiplier), NA, NA, stats::var(used_trait_axis$surv_dev),
+        stats::var(used_trait_axis$obs_dev), stats::var(used_trait_axis$size_dev),
+        stats::var(used_trait_axis$sizeb_dev), stats::var(used_trait_axis$sizec_dev),
+        stats::var(used_trait_axis$repst_dev), stats::var(used_trait_axis$fec_dev),
+        stats::var(used_trait_axis$jsurv_dev), stats::var(used_trait_axis$jobs_dev),
+        stats::var(used_trait_axis$jsize_dev), stats::var(used_trait_axis$jsizeb_dev),
+        stats::var(used_trait_axis$jsizec_dev), stats::var(used_trait_axis$jrepst_dev),
+        stats::var(used_trait_axis$jmatst_dev), stats::var(used_trait_axis$indcova),
+        stats::var(used_trait_axis$indcovb), stats::var(used_trait_axis$indcovc))
+      
+      likely_candidates <- which(var_vector != 0)
+      
+      if (length(likely_candidates) > 0) {
+        best_candidate <- likely_candidates[1]
+        
+        all_possibilities <- c("Given Rate Value", "Additive Offset Value",
+          "Multiplier Value", NA, NA, "Deviation to Intercept of Survival",
+          "Deviation to Intercept of Observation", "Deviation to Intercept of Primary Size",
+          "Deviation to Intercept of Secondary Size", "Deviation to Intercept of Tertiary Size",
+          "Deviation to Intercept of Reproductive Status", "Deviation to Intercept of Fecundity",
+          "Deviation to Intercept of Juvenile Survival", "Deviation to Intercept of Juvenile Observation",
+          "Deviation to Intercept of Juvenile Primary Size", "Deviation to Intercept of Juvenile Secondary Size",
+          "Deviation to Intercept of Juvenile Tertiary Size", "Deviation to Intercept of Juvenile Reproductive Status",
+          "Deviation to Intercept of Maturity Status", "Deviation to Individual Covariate A",
+          "Deviation to Individual Covariate B", "Deviation to Individual Covariate C")
+        
+        used_trait <- all_possibilities[best_candidate]
+        
+        used_trait_spectrum <- used_trait_axis[, best_candidate + 11]
+      }
+      
+      if (!is.element("xlab", names(further_args))) {
+        found_length_of_range <- length(used_trait_spectrum)
+        
+        if (found_length_of_range < auto_axis_ticks) auto_axis_ticks <- found_length_of_range
+        
+        range_of_spectrum <- round(seq(from = min(used_trait_spectrum),
+          to = max(used_trait_spectrum), length.out = auto_axis_ticks),
+          digits = 3)
+        
+        at_range_of_spectrum <- seq(from = 1, to = found_length_of_range,
+          length.out = auto_axis_ticks)
+        
+        further_args$xaxt = "n"
+        further_args$xlab <- used_trait
+        elast_args$side <- 1
+        elast_args$at <- at_range_of_spectrum
+        elast_args$labels <- range_of_spectrum
+      }
+    }
+      
     c_xy <- xy.coords(x = c(1:length(used_fitness)), y = used_fitness)
     further_args$x <- c_xy
     
     do.call("plot.default", further_args)
+    if (auto_axis[1]) do.call("axis", elast_args)
     
     basal_args$x <- c(1:length(used_fitness))
     basal_args$y <- rep(0, length(used_fitness))